[ashell] Move common init code to zjs_modules.c for main/ashell
